Move NotNullByDefault to its upper package in order to reduce its visibility
