+++ src/main/java/io/netty/incubator/codec/http3/package-info.java
@@ -21,4 +21,3 @@
 @NotNullByDefault
 package io.netty.incubator.codec.http3;
 
-import io.netty.incubator.codec.http3.internal.NotNullByDefault;
